Add segment file
We no longer erase the complete device, but add a segment file,
containing the segment summary for the complete device.  mklogfs on a
large device now runs fast. :)
However, mklogfs on a large device is still buggy because we don't
handle indirect blocks in the segment file correctly yet.  Coming up
soon...