Retry write operation when interrupted
While creating the file system, a meta-data is written on to the
device. It is possible that a write operation may not write the
complete data and return after a partial write. Typically this happens
when a signal is received.

This fix adds a safe write call, which starts the write operation in
it is interrupted by a signal.

Signed-off-by: Prasad Joshi <prasadjoshi.linux@gmail.com>